+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v0.2.3
+### Added or Changed
+- changed options afterOnTheRight with ltr true|false
+- added viewchanged event (fired when the before or after image is 70 percent or more visible)
+
+## v0.2.2
+### Added or Changed
+- remove toggle-button & description options in favour of events
+
 ## v0.2.1
 - performance update
 - bug fixing